Allow broadcast errors to surface later.
We are currently checking for `isconcretetype`, which is too limiting, as
e.g. CUDA.jl supports isbits-union arrays. Checking for `allocatedinline`
would be better, but let's just do away with the check entirely and have
the array constructor fail in the presence of unsupported element types.
This is better anyway, as some back-ends may not support isbits-unions.

We do however still check for Union{}, as that isn't allocated inline
so would fail array construction. By using Nothing there instead, we give
the GPU kernel (which is expected to throw an error) the chance to execute
and report an exeception dynamically. This also makes it possible to
trace a broadcast invocation under, e.g., Cthulhu.